CrySystem: fix Linux build for CrySystem. Replacements: itoa->sprintf,
_fstat64->fstat64, 3 int->unsigned int. Disable HTTPDownloader.h in
Linux and few other small fixes.

+++ SourceCode/CrySystem/SystemRender.cpp
@@ -883,7 +883,7 @@ void CSystem::UpdateLoadingScreen()
 {
 	if (!m_bEditor)
 	{
-		if (GetIRenderer()->EF_Query(EFQ_RecurseLevel) <= 0)
+		if ((int)GetIRenderer()->EF_Query(EFQ_RecurseLevel) <= 0)
 		{
 			RenderBegin();
 			GetIConsole()->Draw();
